Add s3-file-field
The default `Image` model now uses an `S3FileField` to store the file.

The `README.md` is still missing one step, creating the `django-storage`
bucket in the Minio instance. This needs to be either added to the set
up instructions (which would be a rather obnoxious step during initial
setup) or the `girder/minio-nonroot` image should include that bucket by
default.