Make test factory registration less fragile
This removes the use of pytest_factorboy's @factory decorator as a matter of
policy, as it's too fragile. Factories must be explicitly imported (if they
are defined elsewhere) and registered in conftest.py.

The register function / decorator works by defining a new fixture function
into the local Python file namespace. Pytest only automatically detects and
registers global fixtures if they are within the namespace of conftest.py;
fixtures defined as part of an import of other files are not registered.