Upgrade to django-composed-configuration v0.10.0
django-composed-configuration now provides many requirements (via extras), so
packages which are not referenced directly have been removed as explicit
requirements.

This removes drf-extensions entirely, as that project is broken on the latest
version of DRF, and no longer seems to be issuing releases.

This switches back to drf-yasg, as the project is unforked.