Revert attempt to use UV_ENV_FILE
When using Tox, it may be the case that commands are not invoked with
"uv run", so the envfile isn't getting loaded for commands run via Tox.

Additionally, integration with VSCode and Pytest will not be run via "uv run",
so we can't rely on uv always getting the chance to set environment variables.